Add blocking parent with child row deletion in backend

+++ backend/authors/delete_author.php
@@ -3,6 +3,7 @@
 header('Content-Type: application/json');
 require_once  __DIR__ .  '/../../config/database.php';
 require_once  __DIR__ . '/../../config/helpers.php';
+require_once  __DIR__ . '/../validators/author_db_validators.php';
 
 $author_id = $_POST["id"] ?? null;
 
@@ -21,6 +22,16 @@ if($validation_result['valid'] === false)
 // sanitize author id
 $DB_author_id = trim($author_id);
 $DB_author_id = (int) $DB_author_id;
+
+$author_has_books_validation = DB_validate_author_has_books($conn , $DB_author_id);
+if(!$author_has_books_validation['success'])
+{
+    echo json_encode([
+        'success' => false,
+        'message' => $author_has_books_validation['message']
+    ]);
+    exit;
+}
 
 $query = "DELETE FROM authors WHERE id = ?";
 $stmt = $conn->prepare($query);
